Finetuning code refactor (#1081)
